(#1) Integrate Middleware Middleware is now an active layer in the system, replacing the lone Transport class that existed before.
Middleware is configured at the Agency class level through `Agency::use_middleware <class>, <args...>`. When creating an Agency instance, new instances of each Middleware class will be made and attached to the agency. Whenever an event occurs, the agency will proxy that event through its middleware instances in the order that they were specified.

Currently, all middlewares are idempotent. That is, each middleware instance will get the original copy of data from the agency, rather than the result of the previous middleware. This means they currently work more as a flat layer than a growing stack, which should change before merging this into master.

Key change: ObjectManager is no longer responsible for keeping track of transport layers. Instead, it gets a reference to the agency it belongs to and proxies events through that.