Use "doublemap_api" gem to handle DoubleMap source logic. Almost everything can be done in the source, with one exception being the handling of `Station#stop_code`. The current solution is a monkey patch on the `DoubleMap::Stop` model to generate this attribute for use here.
DoubleMap doesn't explicitly provide this option; in fact, it seems to be pretty unique to CityBus. Because of that, it should not be the responsibility of the gem to provide it. Instead, there should be a way for the primary attribute to be configurable as a proc, which in this case would evaluate the stop_code from the name. This abstracts the needs of this application from the features of the gem.