Perform 2 initial updates when instantiating an Agency. The first update ensures that all data exists when it is needed. The second ensures that all referenced objects exist, using the data that was ensured by the previous update.
If only one initial update is performed, associations between objects can not always be created. For example, when a Route is trying to associate to Stations, those Stations will only exist if the Station Manager's update occurred before the Route Manager's. However, if that is the case, Stations will not be able to associate to the Routes since the Routes will only exist if the Route Manager's update occurs first. Thus, the second update is necessary.