Dealing with #5: Embed Stations in Routes Culmination of the previous 3 commits.
Route events in Conductor are now a little more DRY as well as succinct, thanks to the safe navigation operator (&.) of Ruby 2.3.

Embedding uses the safe navigation to only embed existing station objects, and only occurs on `update` and `activate` events, since there is no real need for them in the `deactivate` event.

Also, Ruby 2.3 is nice&.safe.