Distinguish "included" and "embedded" in serialization. Previously, all attributes were "embedded" during serialization, meaning that if an Object attribute was included, it would always be embedded.
Now, attributes can be "included" and/or "embedded". Attributes that are only "included" will be left as they are. For regular attributes, this means nothing; but for Object attributes, this means that they will be left as identifiers. Object attributes that are also "embedded" will be transformed into hashes of the attributes specified by `nested_serialized_attributes` and `nested_embedded_attributes`. Attributes that are to be embedded *must also* be included. Embedding does not affect non-Object attributes.