Start of spec for Serializable Relatively easy to write, but an interesting issue with Configurations has come up.
As fixed in the previous commit, `Object.configuration.schema` is now delaying it's evaluation of value aliases until `validate` is called for the first time, however, it seems to be called before "spec/serializable_spec.rb" is run while testing. When run alone, the tests pass; but when run as the whole suite, the tests fail, because the configuration for the dummy class (being copied from Object) already has the transformed configuration values in place, so the testing ones do not take effect.

This can and should be fixed by either returning new objects through `validate`, or ensuring that when inheriting configurations, the original values are copied, rather than the transformed ones.