Bundle update. `ruby_wamp_client` was complaining that it couldn't use `json v2.0.2`. Tried a bundle update, but it didn't work. Uninstalling `json 2.0.2` worked fine.
Bundler is apparently going to add a `force_version:` option for gems, which would avoid issues where having a newer version of `json` installed causes the application to fail. However, that feature has been in deliberation for more than 2 years, so there's no real timeline for when the feature will be usable.