Create `shark` executable, start building runtime manager. IMPORTANT: This commit breaks the old execution path of yore. Instead of running `ruby shark.rb`, instead run `bin/shark -c config/citybus.rb`. Run `bin/shark -h` for other options.
`Runner` (the runtime manager) is responsible for preparing the run environment, loading configurations, and starting the agency services. It is a generalization of the previous `shark.rb` into a more configurable system. (Thinking about it, `Runner` should `include Configurable`...)

The reason for these changes, other than for configurability, is to accomodate what will become the `test` environment, capturing all events and providing a REPL to interact with them. See #28.