Rename `config/citybus.rb` to `config/agency.rb`. This is primarily to simplify running the system (you no longer need to specify the `-c` option).
This is also somewhat of a step towards treating the system as a gem, rather than an application. The idea is to have the system internals kept in one file location (i.e., as an installed gem), and then multiple applications (each in their own directory) can use the system with their own configurations, sources, middlewares, etc.

As of now, it would be possible to create a Shark gem, but I'm going to hold off as the system is still being developed alongside the CityBus application part.