Initial work on better environments Environments are now first-class citizens. In other words, there is an `Environment` class that can be subclassed, and those subclasses can register under different names. The `--environment` option of the CLI then references the list of registered environments to determine which environment to execute in. Included by default are Development, Test, and Production. Production will be used if no argument is given at the command line (this should probably be changed).
Additionally, this commit abstracts Middleware management out of the Agency and into the Environment. This allows different Environment classes to modify the middleware stack more easily. For example, `TestEnvironment` removes any `Transport` middlewares and inserts the `Test::REPL` middleware via a `before(:finalize)` callback. The current list of callbacks include `configure` (where the configuration files get loaded), `load` (where the agency and initial, configured middlewares are created), `finalize` (where the middleware list is converted into an actual stack), and `start` (where the agency schedules it's managers and begins publishing data).

Creating a new Environment is as simple as adding a .rb file in the `environments/` directory, subclassing `Shark::Environment`, and specifying a `cli_name` used to reference it on the command line. For example:

```
  # in ./environments/new_env.rb
  class NewEnv < Shark::Environment
    cli_name :new_env

    before(:load){ puts "In the new environment!" }
  end

  # At the command line
  $ bin/shark -e new_env
  In the new environment!
```