Closes #29. Flush API caches on each update cycle. The issue wasn't that the HTTP requests were getting cached, but that the API library (that I wrote, and should have remembered) uses `memoize` on most of its methods, and the Sources here never tried to flush those local caches.
To fix this, the `refresh` method for each source now includes an `api.flush_cache` call, which should ensure that a new request is made for every cycle.

Note that this solution will hopefully change with the actual fix mentioned in propershark/doublemap_api#1, where there is a global configuration option for caching.